correct cpp07

+++ cpp07/ex01/iter.hpp
@@ -6,7 +6,7 @@
 /*   By: romukena <romukena@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/03/10 17:36:51 by romukena          #+#    #+#             */
-/*   Updated: 2026/03/20 19:00:40 by romukena         ###   ########.fr       */
+/*   Updated: 2026/03/20 19:21:04 by romukena         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,20 +15,20 @@
 #include <iostream>
 
 template <typename T>
-void iter(T *tab, const int size, void (*f)(T &))
+void iter(T *tab, const size_t size, void (*f)(T &))
 {
 	if (!tab || !f || size <= 0)
 		return;
-	for (int i = 0; i < size; i++)
+	for (size_t i = 0; i < size; i++)
 		f(tab[i]);
 };
 
 template <typename T>
-void iter(const T *tab, const int size, void (*f)(const T &))
+void iter(const T *tab, const size_t size, void (*f)(const T &))
 {
 	if (!tab || !f || size <= 0)
 		return;
-	for (int i = 0; i < size; ++i)
+	for (size_t i = 0; i < size; ++i)
 		f(tab[i]);
 }
